# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface User {
  id: string;
  username: string;
  name: string;
  role: string;
  shift: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserInput {
  username: string;
  password: string;
  name: string;
  role: string;
  shift?: string;
}

export interface UserUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  shift?: string | null;
  /** @nullable */
  password?: string | null;
}

export interface Table {
  id: string;
  name: string;
  type: string;
  status: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  openedById?: string | null;
  /** @nullable */
  openedByName?: string | null;
  createdAt: string;
}

export interface TableInput {
  name: string;
  type: string;
}

export interface TableUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  type?: string | null;
}

export interface Session {
  id: string;
  tableId: string;
  tableName: string;
  tableType: string;
  openedById: string;
  openedByName: string;
  /** @nullable */
  closedById?: string | null;
  /** @nullable */
  closedByName?: string | null;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  amount?: number | null;
  pricePerHour: number;
  date: string;
  syncedToSheets: boolean;
  createdAt: string;
}

export interface SessionOpenInput {
  tableId: string;
  userId: string;
  pricePerHour: number;
}

export interface SessionOpenResult {
  success: boolean;
  table: Table;
}

export interface SessionCloseInput {
  tableId: string;
  userId: string;
}

export interface SessionCloseResult {
  success: boolean;
  session: Session;
  table: Table;
}

export interface Setting {
  key: string;
  value: string;
}

export interface SettingInput {
  value: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface DashboardSummary {
  activeTablesCount: number;
  totalTablesCount: number;
  todayRevenue: number;
  todaySessionsCount: number;
  todayTotalMinutes: number;
}

export type GetSessionsParams = {
/**
 * @nullable
 */
date?: string | null;
/**
 * @nullable
 */
userId?: string | null;
/**
 * @nullable
 */
tableId?: string | null;
};