# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Billiards management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: tables
    description: Billiards table management
  - name: sessions
    description: Table sessions (open/close)
  - name: settings
    description: App settings
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /tables:
    get:
      operationId: getTables
      tags: [tables]
      summary: Get all tables
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a new table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"

  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get a table by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /sessions/open:
    post:
      operationId: openSession
      tags: [sessions]
      summary: Open a table session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionOpenInput"
      responses:
        "200":
          description: Session opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionOpenResult"

  /sessions/close:
    post:
      operationId: closeSession
      tags: [sessions]
      summary: Close a table session and calculate bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCloseInput"
      responses:
        "200":
          description: Session closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCloseResult"

  /sessions:
    get:
      operationId: getSessions
      tags: [sessions]
      summary: Get session history with optional filters
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: userId
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: tableId
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setting"

  /settings/{key}:
    put:
      operationId: upsertSetting
      tags: [settings]
      summary: Upsert a setting by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingInput"
      responses:
        "200":
          description: Updated setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"

  /users:
    get:
      operationId: getUsers
      tags: [auth]
      summary: Get all users (manager only)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [auth]
      summary: Create a new user (manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    patch:
      operationId: updateUser
      tags: [auth]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [auth]
      summary: Delete a user (manager only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, username, name, role, shift, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        role:
          type: string
        shift:
          type: string
        createdAt:
          type: string

    UserInput:
      type: object
      required: [username, password, name, role]
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
        shift:
          type: string

    UserUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        shift:
          type: ["string", "null"]
        password:
          type: ["string", "null"]

    Table:
      type: object
      required: [id, name, type, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startTime:
          type: ["string", "null"]
        openedById:
          type: ["string", "null"]
        openedByName:
          type: ["string", "null"]
        createdAt:
          type: string

    TableInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string

    TableUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]

    Session:
      type: object
      required: [id, tableId, tableName, tableType, openedById, openedByName, startTime, pricePerHour, date, syncedToSheets, createdAt]
      properties:
        id:
          type: string
        tableId:
          type: string
        tableName:
          type: string
        tableType:
          type: string
        openedById:
          type: string
        openedByName:
          type: string
        closedById:
          type: ["string", "null"]
        closedByName:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        amount:
          type: ["integer", "null"]
        pricePerHour:
          type: integer
        date:
          type: string
        syncedToSheets:
          type: boolean
        createdAt:
          type: string

    SessionOpenInput:
      type: object
      required: [tableId, userId, pricePerHour]
      properties:
        tableId:
          type: string
        userId:
          type: string
        pricePerHour:
          type: integer

    SessionOpenResult:
      type: object
      required: [success, table]
      properties:
        success:
          type: boolean
        table:
          $ref: "#/components/schemas/Table"

    SessionCloseInput:
      type: object
      required: [tableId, userId]
      properties:
        tableId:
          type: string
        userId:
          type: string

    SessionCloseResult:
      type: object
      required: [success, session, table]
      properties:
        success:
          type: boolean
        session:
          $ref: "#/components/schemas/Session"
        table:
          $ref: "#/components/schemas/Table"

    Setting:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
        value:
          type: string

    SettingInput:
      type: object
      required: [value]
      properties:
        value:
          type: string

    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    DashboardSummary:
      type: object
      required: [activeTablesCount, totalTablesCount, todayRevenue, todaySessionsCount, todayTotalMinutes]
      properties:
        activeTablesCount:
          type: integer
        totalTablesCount:
          type: integer
        todayRevenue:
          type: integer
        todaySessionsCount:
          type: integer
        todayTotalMinutes:
          type: integer