# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  activeTablesCount: number;
  totalTablesCount: number;
  todayRevenue: number;
  todaySessionsCount: number;
  todayTotalMinutes: number;
}