# lib/api-zod/src/generated/types/sessionCloseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionCloseInput {
  tableId: string;
  userId: string;
}