# lib/api-zod/src/generated/types/sessionCloseResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */
import type { Session } from './session';
import type { Table } from './table';

export interface SessionCloseResult {
  success: boolean;
  session: Session;
  table: Table;
}