# lib/api-zod/src/generated/types/sessionOpenResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */
import type { Table } from './table';

export interface SessionOpenResult {
  success: boolean;
  table: Table;
}