# lib/api-zod/src/generated/types/table.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */

export interface Table {
  id: string;
  name: string;
  type: string;
  status: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  openedById?: string | null;
  /** @nullable */
  openedByName?: string | null;
  createdAt: string;
}