# lib/api-zod/src/generated/types/tableUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */

export interface TableUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  type?: string | null;
}