# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  username: string;
  name: string;
  role: string;
  shift: string;
  createdAt: string;
}