# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billiards management API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  shift?: string | null;
  /** @nullable */
  password?: string | null;
}